+++ cardano-testnet/CHANGELOG.md
@@ -7,6 +7,7 @@
 * [Add flag to support P2P topology](https://github.com/IntersectMBO/cardano-node/pull/6263)
 * [Add flag to update time stamps in custom environment](https://github.com/IntersectMBO/cardano-node/pull/6275)
 * [Add option to create configuration sandbox with parameters from mainnet](https://github.com/IntersectMBO/cardano-node/pull/6289)
+* [Remove deprecated options from CLI](https://github.com/IntersectMBO/cardano-node/pull/6314)
 
 ## 10.0.0